requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
anthropic==0.43.0
pydantic==2.10.0
python-docx==1.1.2
weasyprint==63.1
python-dotenv==1.0.1
